+++ api/pkg/provider/cloud/provider.go
@@ -8,7 +8,7 @@ import (
 	"github.com/kubermatic/kubermatic/api/pkg/provider/cloud/fake"
 	"github.com/kubermatic/kubermatic/api/pkg/provider/cloud/hetzner"
 	"github.com/kubermatic/kubermatic/api/pkg/provider/cloud/openstack"
-	"github.com/kubermatic/kubermatic/api/pkg/provider/cloud/vmsphere"
+	"github.com/kubermatic/kubermatic/api/pkg/provider/cloud/vsphere"
 )
 
 // Providers returns a map from cloud provider id to the actual provider.
