+++ api/pkg/provider/cloud/vsphere/helper.go
